VEC-Scheduler-App/flask-server/requirements.txt
flask
python-dotenv
flask_sqlalchemy
flask-bcrypt
flask-sessions
flask_cors
flask_session
pymysql
python-dotenv

